Crate re_data_source

Handles different ways of loading Rerun data, e.g.:

Over HTTPS

Over WebSockets

From disk

Also handles different file types: rrd, images, text files, 3D models, point clouds…

Modules

data_loader 🔒

data_source 🔒

load_file 🔒

load_stdin 🔒

web_sockets 🔒

Structs

ArchetypeLoader

Loads data from any supported file or in-memory contents as native re_types::Archetypes.

DataLoaderSettings

Recommended settings for the DataLoader.

DirectoryLoader

Recursively oads entire directories, using the appropriate crate::DataLoader:s for each files within.

ExternalLoader

A crate::DataLoader that forwards the path to load to all executables present in the user’s PATH with a name that starts with EXTERNAL_DATA_LOADER_PREFIX.

FileContents

The contents of a file.

RrdLoader

Loads data from any rrd file or in-memory contents.

Enums

DataLoaderError

Errors that might happen when loading data through a DataLoader.

DataSource

Somewhere we can get Rerun data from.

LoadedData

What DataLoaders load.

Constants

EXTERNAL_DATA_LOADER_INCOMPATIBLE_EXIT_CODE

When an external crate::DataLoader is asked to load some data that it doesn’t know how to load, it should exit with this exit code.

EXTERNAL_DATA_LOADER_PREFIX

To register a new external data loader, simply add an executable in your $PATH whose name starts with this prefix.

SUPPORTED_IMAGE_EXTENSIONS

SUPPORTED_MESH_EXTENSIONS

SUPPORTED_POINT_CLOUD_EXTENSIONS

SUPPORTED_RERUN_EXTENSIONS

SUPPORTED_TEXT_EXTENSIONS

Traits

DataLoader

A DataLoader loads data from a file path and/or a file’s contents.

Functions

connect_to_ws_url

on_msg can be used to wake up the UI thread on Wasm.

extension

Empty string if no extension.

is_supported_file_extension

Is this a supported file extension by any of our builtin DataLoaders?

iter_external_loaders

Iterator over all registered external crate::DataLoaders.

iter_loaders

Iterator over all registered DataLoaders.

load_from_file_contents

Loads the given contents using all crate::DataLoaders available.

load_from_path

Loads the given path using all crate::DataLoaders available.

register_custom_data_loader

Register a custom DataLoader.

supported_extensions

All file extension supported by our builtin DataLoaders.