Function re_data_source::load_stdin::load_stdin

pub fn load_stdin(tx: Sender<LogMsg>) -> Result<()>

Asynchronously loads RRD data streaming in from standard input.

This fails synchronously iff the standard input stream could not be opened, otherwise errors are handled asynchronously (as in: they’re logged).